Fix password input
While the password input is required, the provided value wasn't actually used. Instead, the password was set to whatever was in the TEST_NEO4J_PASSWORD environment variable. While this happened to work fine for all initial users of this action, that's obviously a bug.

With this change, the environment variable is ignored and the password is set to the input value.